refactor:컴포넌트 분리 및 AuthEorrorModal 추가-#116
